requirements.txt
tkinterdnd2
reportlab
Pillow
docx2pdf